+++ frontend/src/components/common/inputs/AsyncAutocomplete.vue
@@ -2,20 +2,27 @@
   <v-autocomplete
     v-bind="{ ...$props, ...$attrs }"
     :items="options"
-    :loading="loading"
+    :loading="isLoading"
     :filter="filter"
     :search-input.sync="searchInput"
     hide-no-data
     v-on="$listeners"
   >
     <template #append>
-      <v-icon>mdi-database-search-outline</v-icon>
+      <v-tooltip bottom>
+        <template #activator="{ on, attrs }">
+          <v-icon v-bind="attrs" class="db-icon" v-on="on">mdi-database-search-outline</v-icon>
+        </template>
+        <span> Realiza búsqueda en BD </span>
+      </v-tooltip>
     </template>
   </v-autocomplete>
 </template>
 
 <script>
 import { debounce, isObject } from "lodash";
+
+import useLoading from "@/composables/useLoading";
 
 export default {
   name: "AsyncAutocomplete",
@@ -66,10 +73,18 @@ export default {
       default: "Escribe para iniciar la búsqueda...",
     },
   },
+  setup() {
+    const { isLoading, addTask, removeTask, isTaskLoading } = useLoading();
+    return {
+      isLoading,
+      addTask,
+      removeTask,
+      isTaskLoading,
+    };
+  },
   data() {
     return {
       options: [],
-      loading: false,
       searchInput: null,
     };
   },
@@ -87,7 +102,7 @@ export default {
             if (isObject(newValue)) {
               this.options = [newValue];
             } else {
-              this.loading = true;
+              this.addTask("fetch-item");
               try {
                 const response = await this.service[this.getFunctionName](newValueId);
                 this.options = [response.data];
@@ -95,7 +110,7 @@ export default {
                   this.$emit("input", response.data);
                 }
               } finally {
-                this.loading = false;
+                this.removeTask("fetch-item");
               }
             }
           }
@@ -105,7 +120,7 @@ export default {
     },
     searchInput(newValue) {
       if (newValue) {
-        if (this.loading) {
+        if (this.isTaskLoading("search")) {
           return;
         }
         const exactMatch = this.options.find((item) => newValue.toLowerCase() === item[this.itemText].toLowerCase());
@@ -119,8 +134,7 @@ export default {
   },
   methods: {
     async search(query) {
-      this.loading = true;
-
+      this.addTask("search");
       try {
         const requestParams = { page_size: 15, page: 1 };
         const lookup = this.searchLookup ? `__${this.searchLookup}` : "";
@@ -130,7 +144,7 @@ export default {
         const response = await this.service[this.searchFunctionName](requestParams);
         this.options = response.data.results;
       } finally {
-        this.loading = false;
+        this.removeTask("search");
       }
     },
     debouncedSearch: debounce(function (query) {
@@ -147,3 +161,10 @@ export default {
   },
 };
 </script>
+
+<style scoped>
+.db-icon {
+  cursor: help;
+  opacity: 0.6;
+}
+</style>