+++ frontend/src/components/layout/AppNavbar.vue
@@ -7,16 +7,15 @@
       <img v-if="isXmas" id="santaHat" src="@/assets/santa_hat.svg" alt="Santa's hat" />
     </span>
     <v-spacer />
-    <v-progress-circular v-show="loading" class="mr-5" :size="36" color="white" indeterminate />
+    <v-progress-circular v-show="loadingRequests" class="mr-5" :size="32" color="white" indeterminate />
     <LinkManager />
     <NotificationManager class="mr-5" />
+    <UserAvatar size="36" :font-size="14" :user="loggedUser" />
     <v-menu bottom left offset-y>
       <template #activator="{ on, attrs }">
-        <span id="userMenu" v-bind="attrs" v-on="on">
-          <span class="mr-3">{{ loggedUser.first_name }}</span>
-          <UserAvatar size="36" :font-size="14" :user="loggedUser" />
-          <v-icon right>mdi-menu-down</v-icon>
-        </span>
+        <v-btn icon v-bind="attrs" v-on="on">
+          <v-icon>mdi-dots-vertical</v-icon>
+        </v-btn>
       </template>
       <v-list>
         <v-list-item @click="$vuetify.theme.dark = !$vuetify.theme.dark">
@@ -51,7 +50,7 @@ export default {
   components: { LinkManager, NotificationManager },
   computed: {
     ...mapState(["loggedUser"]),
-    ...mapGetters(["appLabel", "loading"]),
+    ...mapGetters(["appLabel", "loadingRequests"]),
     isXmas,
   },
   methods: {
@@ -80,8 +79,5 @@ export default {
   position: absolute;
   left: 2.73em;
   top: 0.4em;
-}
-#userMenu {
-  cursor: pointer;
 }
 </style>