+++ frontend/src/components/scrum/UserStoryStatus.vue
@@ -1,10 +1,15 @@
 <template>
-  <v-stepper class="elevation-0" :value="userStory.status">
+  <v-stepper :value="userStory.status" flat>
     <v-stepper-header>
       <template v-for="(status, index) in userStoryStatus">
-        <v-stepper-step :key="status.value" :step="status.value" :complete="userStory.status > status.value">{{
-          status.label
-        }}</v-stepper-step>
+        <v-stepper-step
+          :key="status.value"
+          :step="status.value"
+          :complete="userStory.status > status.value"
+          :color="userStory.status > status.value ? 'green' : 'primary'"
+        >
+          {{ status.label }}
+        </v-stepper-step>
         <v-divider v-if="index < userStoryStatus.length - 1" :key="status.label" />
       </template>
     </v-stepper-header>