+++ frontend/src/components/scrum/filters/UserStoryFilters.vue
@@ -45,9 +45,7 @@
         />
       </v-col>
       <v-col lg="2" sm="6" cols="12">
-        <v-btn class="my-2" color="primary" :loading="loading" :disabled="loading" @click="$emit('apply:filters')">
-          Filtrar
-        </v-btn>
+        <v-btn class="my-2" color="primary" :disabled="disabled" @click="$emit('apply:filters')"> Filtrar </v-btn>
       </v-col>
     </v-row>
 
@@ -426,7 +424,7 @@
           <v-btn color="warning" text @click="clearFilters">Limpiar</v-btn>
           <v-spacer />
           <v-btn text @click="closeFiltersDialog"> Volver </v-btn>
-          <v-btn text :loading="loading" :disabled="loading" @click="applyFiltersFromDialog"> Filtrar </v-btn>
+          <v-btn text :disabled="disabled" @click="applyFiltersFromDialog"> Filtrar </v-btn>
         </v-card-actions>
       </v-card>
     </v-dialog>