+++ frontend/src/main.js
@@ -1,7 +1,6 @@
 import Vue from "vue";
-import VueCompositionAPI from "@vue/composition-api";
 
-Vue.use(VueCompositionAPI);
+import "./plugins/composition";
 
 import router from "./router";
 import store from "./store/index";