+++ frontend/vue.config.js
@@ -1,11 +1,11 @@
 const BundleTracker = require("webpack-bundle-tracker");
 
 module.exports = {
-  publicPath: process.env.NODE_ENV === "production" ? "/static/" : "http://localhost:8080/",
-  transpileDependencies: ["vuetify"],
+  publicPath: process.env.NODE_ENV === "development" ? "http://localhost:8080/" : "/static/",
   devServer: {
     host: "0.0.0.0",
     port: 8080,
+    public: "0.0.0.0:8080",
     https: false,
     headers: { "Access-Control-Allow-Origin": ["*"] },
     hotOnly: true,
@@ -14,6 +14,10 @@ module.exports = {
       aggregateTimeout: 300,
       poll: 1000,
     },
+  },
+  transpileDependencies: ["vuetify"],
+  css: {
+    sourceMap: true,
   },
   chainWebpack: (config) => {
     config.plugin("BundleTracker").use(BundleTracker, [